Hbaseutils TimeUtils UserState 修改了以上三个

+++ src/main/scala/cn/itcast/pyg/realprocess/bean/Message.scala
@@ -1,7 +1,7 @@
 package cn.itcast.pyg.realprocess.bean
 
 case class Message (
-                     var count:Int = 0,
+                     var count:Long = 0,
                      var timeStamp: Long = 0L,
                      var userScan: UserScan = null
                    )
